Make it so exiting with changes will warn
With these changes, typing in 'exit' while there are unsaved changes
will initiate a prompt that the user has to answer 'y' to in order to
exit. If they answer anything else, the session does not exit.

Additionally, KeyboardInterrupt will now show whether we exited when
there were unsaved changes.