Improve compatibility with insane filenames.
This allows us to install mods with quotes in the filenames of the
downloads. We do this by renaming the download into one with a better
name. This only happens if we try to install it. We have to do this to
avoid issues with our os.system(7z) call.