Prevent more false alarms on redundant dirs
We trim redundant dirs in cases of mods being set up like this:
modname/redundant_folder/Data. In this case, Data would be moved into
the modname folder. This logic was causing issues for mods that only
contained folders normally found in Data, but not plugins. The main
folders found in Data have been added to a list that is checked against
to prevent elevating folders that shouldn't be elevated.