Improve tolerance to existing files
Ammo will no longer crash when attempting to create a symlink on top of
an existing file. Instead, it will print a message saying the file
existed so creation of the link will be skipped.

This can happen if you attempt to install a mod that overwrites default
assets, or assets that were moved into the game dir via a method other
than Ammo.