Prevent crashes from extracting/deleting certain filetypes
Attempting to extract a _folder_ that ended in an _archive_ extension
like .7z would cause no files to be extracted, which would crash Ammo.
There was a similar issue for deleting directories that incorrectly had
archive extensions.

Now an error message is printed instead that nothing was extracted, or
that Ammo refuses to delete archives.