Only show plugins found in loadable locations
Make plugins only populate in a mod's owned plugins list if the plugin
file was found in the root folder of the mod or in mod_root/Data.

Previously, plugins found anywhere within the mod's directories were
being added. It was possible to add plugins to Plugins.txt that couldn't
be found by the game to load.