Prevent polluting Data dir when bailing from fomod install
There were some conditions that caused bailing on a fomod install to
copy files from the mod's dir to the data dir (as opposed to copying
files from the mod's data dir to the data dir). Prevent this by forcing
fomods to be disabled when you try to reconfigure them.