Make fomod installer handle nested files
When a ModuleConfig.xml listed a folder source as a nested directory,
the fomod installer now properly constructs the location.

Previously, the folder source was treated as a single file, like the
literal backslash was part of the filename instead of a
descent-indicator.

This fixes RSS Children Overhaul, and probably also a bunch of other
mods.

This change also makes the fomod installer commands 'next' and 'back'
simply be 'n' and 'b' instead. This is because some fomod installers
have upwards of 20 pages, and it was brutal to type 'next' that many
times.