Respect fomod file source and destination folders
Previously, the destination directive for files in ModuleConfig.xml was
not being handled correctly, causing files to be installed to the wrong
location. Additionally, mods that contained DLLs were auto-configured
even if they were fomods, which caused issues with Powerofthree's mods.

Now, when fomods are configured, the destination directive in
ModuleConfig.xml is respected, so a mod-local Data folder is created and
the chosen components are copied into it with the expected parent paths.

Mods with DLLs that are fomods are no longer auto-configured.

This fixes (at least) the following mods:
- Powerofthree's Tweaks
- Base Object Swapper