Fix fomods overwriting their own files
The shutil.copytree utility will overwrite a dir if it exists instead of
copy files into it. This was causing only the last-in-line file to be
moved into place.

Additionally, store visibility flags so unnecessary fomod steps can be
skipped later. Currently there is no consumer for the visibility flags,
and flipping appropriate flags upon selection change still needs to be
handled properly.