Fix SSE Display Tweaks installation
Previously, the logic that was used to detect whether a DLL style mod
should be installed to the game's directory rather than the game's Data
directory was not double checking whether the DLL existed inside of
SKSE/Plugins. This was causing SSE Display Tweaks to be installed to
Skyrim/SKSE/... instead of Skyrim/Data/SKSE/...

Now, SSE Display Tweaks will install to the proper location.