Include conditional plugins
Previously, only plugins that were unconditional or were in a
conditionalFileInstalls section in ModuleConfig.xml were being included.
Mods like SMIM have ModuleConfig.xml files where the plugins themselves
are conditional but still contain files. Those files are now being
included.

Additionally, they have their source locations corrected for
case-sensitivity, so if a file is declared in ModuleConfig.xml in
lowercase but the actual file on the filesystem is uppercase, it should
now be properly found and included.

With these changes, SMIM should now be largely functional, with the
exception of a missing chair texture.