Fix unconditional file installs
Unconditional file installs stopped functioning due to a recent change.
Lists were not being flattened, which was causing the source location of
files to not be extracted properly. This caused a crash.

With these changes, power of three's tweaks should be working again,
along with any other mods shaped like it.

This change also renames mod.data_dir to mod.has_data_dir so I can stop
confusing it with a location.