Add fuzzy "find" command
The "find" command, issued without arguments, will show all files.
"find" accepts any number of arguments. Any plugin, mod, or download
that contains any of the arguments will be shown, and all others will be
hidden.

This approach is useful for showing a specific mod/plugin, but also
useful for showing a variety, in case you need to view a specific
combination of mods/plugins at once.

This also contains a change for the way partially completed downloads
are identified, and should help reduce false positives on complete
downloads when using Firefox.