Avoid multiple return types
This hits most functions in controller, but not all of them.
_fomod_get_root_node, for example, can still return either False or
ElementTree.

This change reduces repetitive code in the controller by removing
validation and raising errors instead of returning false when validation
failed.

Error messages are now mostly defined in the UI, as each type of error
generally is only created by a specific issue. For example, IndexError
will likely only occur in the cases of a user entering a number out of
range, so IndexError will simply print "Index out of range".

Because of this, error messages are more generic. The simplicity gains
are a worthwhile trade-off.