Change to package format, requires reinstall but not of mods
This commit changes the format of the project to that of a normal python
package. This should make the way you interact with installing and
updating ammo a bit more standard, and also helps by allowing tests to
be run from the clone directory instead of the test directory, which
will allow for easier test automation.

Please see the new update/install instructions in the README. You do not
need to reinstall or reconfigure mods, but you do need to modify a few
things to benefit from this and subsequent updates.