Fix an issue where fomod plugins weren't installing
Some fomods (like Embers XD), have plugins both inside and outside of
the data folder. The way that we were populating fomods
filename:/path/to/filename dict was not accommodating a condition where
a plugin could exist outside of the Data folder, as well as inside of
the data folder.

Previously, when activating Embers XD, a plugin would populate into your
plugins list. You could activate and commit this plugin. Refreshing
would cause that plugin to disappear, as it was never installed.

Fix the logic surrounding fomod file population so plugins like this
correctly persist.