Inherit from abstract Controller
The UI now expects an instance of a Controller, which is an abstract
base class. Our old Controller class has been renamed to ModOrganizer,
and it now also inherits from Controller.

This will allow us to define an interface that all of ModOrganizer, the
UI, and a yet-to-come FomodController can agree on.