Move Fomod-specific logic to FomodController
Refactor the fomod specific logic that was previously in ModController
into FomodController, which allows us to recycle the UI class and
organizes the code a bit better.

Note that to configure fomods, you now have to issue "select <index>"
instead of simply "<index>". Sorry for the inconvenience. I'll think
about ways to make it more convenient as part of future work.