Make the UI support dynamically allocated methods
Previously, the UI was using inspect.getmembers to identify methods to
map to commands, and was only mapping them to commands on init.

getmembers doesn't care about instance members (only class members) so
any dynamically generated methods for the controller wouldn't be mapped.

Destroy and recreate UI.commands on init and at the top of every rep so
dynamic commands would vanish and re-appear.

This functionality will hopefully be able to be used by the
FomodController to enable us to use `<index>` instead of `select
<index>`, like the good 'ol days (like 8 days ago?).