Fix "edit scripts" mods
Fix mods that contain only an "edit scripts" folder at the top level.
Previously AMMO was extracting the contents of "edit scripts" to the
base directory, which makes xEdit unable to find the scripts.

Simply add a condition to check for "edit scripts" so this doesn't
happen. Fixes "Check for Missing Textures Nif Script", which helped me
diagnose the issue AMMO was having with SMIM and RW2.