De-duplicate the normalize function
The normalize function is responsible for making sure folders are all
lowercase except for a list of folders which have their capitalization
replaced with the casing of the list item.

Both the fomod controller and the mod controller required this function
and it was defined in both. Move this function out of controllers into
lib.py.