Fix mod.modconf not being re-assigned on rename
There was an issue when renaming a fomod that caused mod.modconf to not
be re-assigned to the new ModuleConfig.txt location. This was causing
crashes when attempting to configure a fomod after you had renamed it.

Ensure mod.modconf is assigned on init, even if self.fomod or
self.has_data_dir is already set.