Remove obsolete attributes for Download
Previously we couldn't extract downloads if they had weird names. We
were using a Download.sane bool to determine whether the download was
possible to install, which is now unused and shouldn't be needed again.

Remove the attribute Download.sane, which allows us to remove the
post_init dunder method.