Don't auto enable disabled mods
Previously we were trying to auto-enable disabled mods if they were
installed perfectly. We can't adequately test this logic, and it was
probably a bad decision to try to support it in the first place, so
don't.

Remove a shadowed variable that was causing the tests for this to pass
when they shouldn't have, and delete those tests.