Refactor mod init logic to improve speed
This is at the cost of accuracy, as ammo will now only check for folders
whose presence determines whether the mod should be installed in
game.directory or game.data within the surface level of the mod.

This allows us to avoid two instances of os.walk, which was as slow as
it was ugly.

This introduces a match statement, which means python version below 3.10
are no longer compatible. If you are using an older version of python,
please upgrade to 3.10 or later.