Import readline into UI, set completer
This is the first step in getting readline support working. So far, this
should give us a scrollback buffer so we can do things like use the
arrow keys for cursor position and to access previous command history.
We should also be able to use typical keybinds like ctrl+u to clear the
buffer to the left of the cursor, ctrl+k to clear the buffer to the
right of the cursor, and ctrl+l to clear the screen.

The _autocomplete function has been added to controller classes but it
currently just returns None, which results in no completions. This is
currently only slightly better than before, as we no longer inject tabs
when we may have been searching for completion functionality. Tabs are
never useful, so this is fine.

Future work includes adding logic to controller._autocomplete to make it
actually useful.