Add ability to delete plugins
Add the ability to execute commands like:
delete plugin 5
delete plugin all

The help text in the README has also been updated to indicate this
change.

Note that we continue to not want to be able to rename plugins due to
this being able to break your game (if you rename a plugin that another
depended on, for example).