Add the 'sort' command for auto-arranging plugins
Add new 'sort' command that arranges plugin order according to order of
enabled mods that provide those plugins. This is done in reverse to
ensure only conflict winners are considered for ordering.

Add several tests to test_pending_changes to ensure commands either
create a pending change or require an absence of pending changes as
expected.