Make sorting retain user order per-mod
The sort function now retains the order of user's plugins when those
plugins belong to the same mod. For example, if you have one mod that
owns 3 plugins, those plugins will be moved next to each other if they
weren't already, but the first plugin of that mod will still load before
the other two, etc.