Don't support "Data" as a source location for fomods
Fomods that pull files from <fomod_folder>/Data aren't supported since
we have to delete the Data folder when we reconfigure them. Produce a
useful error when this happens to indicate manual intervention is
required.

So far, this impacts new versions of MCM Helper. There might or might
not be others.