Fix multiple bugs related to plugins not populating
The recent patches highlighted the need to refactor the logic that
populates plugins, which I've been encountering bugs with for quite some
time.

component.py::Mod post init logic is responsible for populating that
mod's files and detecting plugins and adding them to that mod's list of
plugins. Previously, we were finding files and plugins at the same time.
This made detecting whether a plugin was in the correct location overly
complicated, which caused plugins to either show up when they weren't
supposed to, or not show up when they were.

Separate the 'populating mod.files' logic from 'populating mod.plugins'.
This allows us to accurately identify the folder we expect plugins to be
in, and only check that folder for plugins. It's still fast since we can
do a shallow peek instead of a recursive walk.