Add experimental custom path support
Ammo will now search in the --conf dir for files ending in .json. The
filename represents the name of a game, and the contents define the
properties. Games defined like this will appear in the game selection
menu and should be possible to manage.

This is primarily written to get GoG up and running, but I'm currently
running into an issue where the GoG version of the game won't load
plugins and overwrites Plugins.txt with a default every time I try to
launch the game.

In any case, see https://github.com/cyberrumor/ammo/issues/42 for an
example of a game configured like this.