Add step name to fomod install wizard
Some mods, like ENB Light, have option names that are the same on
several pages, and are only differentiated via the name of the actual
step. This previously made it look like we were displaying the same page
multiple times even though we weren't.

Make the fomod install wizard present the name of the step (if it
exists) so there is more indication that the installer is working as
intended.