Require absence of changes for tools
When transitioning to the tools menu, require absence of changes. This
is because we need to refresh after leaving the tools menu, but before
returning to the normal mod organizer. If we don't refresh, we could be
looking at downloads that have already been deleted, or missing plugins
that were added by a tool while we were looking at the other menu.