Mod.plugins is a list of Paths instead of strs.
When Mod.plugins was a list of strings, we didn't have enough
information to safely delete plugins without doing a ton of validation
that those plugin files were actually under Data. This validation code
was messy, and I wanted to keep the logic that detects whether a file is
a plugin isolated to Mod.__post_init__.

Make Mod.plugins a list of Paths. This allows us to greatly simplify the
plugin deletion logic and still avoid deleting files we aren't supposed
to.

In removing this validation, the logic that determines whether a file is
appropriately isolated to Mod.__post_init__.