Activate all mods ignores unconfigured fomods
Historically, 'activate mod all' would activate all mods up until it
encountered a warning, at which point it would stop trying to activate
further mods. This was causing self._stage() to not get run, which was
causing issues with the conflict detection system.

Instead of ensuring self._stage() runs even if there's a warning, just
keep attempting to activate the rest of the mods and store warnings.
After this, run self._stage() then display all warnings encountered.

This means all mods will have activation attempted, even if they were
sorted after an unconfigured fomod.

This also fixes the problem of the collision detection system not
working when a warning was raised from this.