Move 'operator' above flags dict
Previously we were tracking plugin dependencies in fomod installers via
a dict. That dict was also responsible for storing the operator (like
'and', 'or', etc). This had the potential to create issues for fomods
that had a flag named 'operator'.

The operator and flags are now stored inside a Dependency dataclass.
They're next to each other rather than nested, so the above problem is
no longer possible.