Optimize lib.normalize
lib.normalize is a pretty hot loop and was (and still is) the vast
majority of ammo's startup and staging time.

The main bottleneck in it was using path.parent, which is slow for
whatever reasons.

```
str(my_path).rsplit("/", 1)[0]
```

achieves the same thing as `my_path.parent`, but it's much faster. So do
that instead.