Decorate methods which require absence of changes
Instead of checking within each function that requires a commit or
refresh before executing that self.changes is False, use a decorator
that does this.

This makes methods which have this requirement visually distinct (when
looking at the code).

This makes the warning that is raised when attempting to execute a
decorated method generic (it no longer includes the name of the action
the user tried to perform).