Indicate Bethesda-specific attributes with comment
Add a few comments to various classes to indicate which attributes are
specific to bethesda games and which are generic. This does not include
Enums (only RenameEnum and ToolEnum are generic).

Reorder the attributes in the Mod class and in the init method of
ModController to put all generic and bethesda attributes closer to each
other on the page.