Make UI expose 'do_' methods instead of public methods
In order to bring ammo closer to standard lib design paradigms (like
that of the cmd module), make the UI class expose Controller methods
prefixed with 'do_' instead of all methods which don't begin with '_'.

Update method names for all controllers so they utilize no underscores
for normal methods, and utilize the 'do_' prefix for methods which
should be exposed to the UI class.