Don't override self in lambdas
There's a few lambda expressions in populate_index_commands which are
used as the command which is executed when you run "0" or similar from
the fomod_controller or game_controller.

Those lambda commands were previously defined as

```
lambda self, i=i: self.some_method(i)
```

This previously worked fine, but when testing a migration to cmd.Cmd (as
opposed to using the UI class), this code bugged out. Somehow, 'self'
had become a string.

If we don't use 'self' in the lambda arguments, this doesn't occur. No
idea why it was working as-is (without the cmd.Cmd stuff).

Prevent the potential breakage by not passing 'self' into the lamdba.

```
lambda _, i=i: self.some_method(i)
```