Reorganize controllers into dedicated folder
Move the controllers into a controller directory. This removes the need
to identify the type of information contained in a particular module by
its filename.

Move bin/ammo to ammo/bin/ammo. This cleans up the base directory of the
repository.

Delete the base level __init__.py. This provides a clear boundary
between the ammo module and tests.