Don't remove extra directories for controller.mod
The generic controller.mod can't assume anything about whether a mod was
packaged directly, because it doesn't have enough information to make
that determination. It must only extract a mod such that the extracted
files can be linked directly under the game directory. It must assume
that all mods are packaged such that they will work when installed like
this.