Add Sims 4 detection to controller.game
Make the game controller detect Sims 4. Set game_selection.directory
(and therefore game.directory) to (inside of the appropriate pfx)
Documents/Electronic Arts/The Sims 4/Mods instead of the _actual_ game
directory. This is because Sims titles expect mods to be installed in
that directory instead of the actual game directory. The game will just
load them from there.

The generic controller.mod has no explicit test coverage associated with
it, so I'm not considering Sims 4 to be supported yet; it's merely
detected.

Manually tested with wicked whims from the nexus, seemed to load up
okay.