Revert "Add Tomb Raider I-III Remastered detection (#66)"
The vast majority of Tomb Raider I-III Remastered mods are asset
replacers that work by overwriting the original game files.

The best way to solve this problem is with a VFS (similar implementation
to MO2). This is outside the scope of ammo since it would require ammo
to launch the game in order to expose the VFS to the game.

This reverts commit d4abbf4192b108960a41ab07bcfd4dca35d86d84.