Rename conflict_x.7z to mock_conflict_x.7z
I was having issues with ModController.has_extra_folder when migrating
tests for test_conflict_resolution (which is handled in an upcoming
patch, not this one).

Instead of making a better solution for that function, just rename the
mods to match the name of the top level folders inside them, which will
make the existing ModController.has_extra_folder behave correctly.