Refactor generic mod installation
In the BethesdaController, we can use tricks like scanning for plugin
files with known extensions to determine whether an extracted mod was
packaged with an extra folder between the extraction directory and the
actual contents that should be installed into the game directory, E.g.,

ammo/<game>/mods/my_mod/my_mod/Data/some_plugin.esp

Here, we know we can remove the extra my_mod folder.

For the generic ModController, there is no cookie-cutter detection
method which can determine whether a folder is unnecessary or not. For
this controller, when there is only one folder in the contents of the
extracted mod directory, always ask the user whether the files in that
folder should be elevated above the folder or not.

This patch also does a bunch of other stuff that needed to happen to
make this possible:

* Make the UI class return controller.return_value if defined and
  controller.postcmd returns True (postcmd's return value determines
  whether we should break from repl).
* Add a BoolPromptController which can be used to return True/False for
  arbitrary questions based on user input.
* Refactor tests, which works around an issue I still don't understand
  where tests were passing in CI but failing locally (even though I
  tried using a venv with the same python version as CI).
* Add ammo_unit_test_env.py - used to pop ammo shell with ModController
  against the unit test game and test downloads folder.
* Add ammo_unit_test_beth_env.py - used to pop ammo shell with
  BethesdaController against the unit test game and test downloads
  folder.
* Use Python 3.12.9 in CI. Python 3.11 is no longer supported because we
  need to use backslashes in f-strings to properly format a question
  used for the BoolPromptController. It's probably possible on an older
  Python version in a different way, but this patch is already massive.