Move folder elevation logic out of extract_archive
I want to cache extracted archives in unit tests so I can copy
pre-extracted archives during installation instead of installing via
extracting archives every single time something needs to be installed
(which results in the same slow extraction step being repeated
unnecessarily).

In order to do this, the extract_archive function needs to extract
without doing any sort of file manipulation after the fact.

There was logic which previously was included in extract_archive which
elevated extracted files out of a nested folder. This was only done when
the nested folder's parent was proven to be safe to remove.

Move this logic out of extract_archive and into do_install. This will
enable caching extracted archives in unit tests to be implemented in the
future.