Remove an instance of os.walk
os.walk is really slow due to I/O, so the less of them we have, the
better.

There is an operation that happens in Mod.__post_init__ which searches
for the directory within a mod which we expected plugin files to be
inside of.

Previously, this operation used os.walk().

Since this operation occurs after we've populated Mod.files, it's safe
to just iterate over our stored files instead of walking the mod again.

Replace this os.walk with an iterator over mod.files.