Cast os.walk dirpath into Path outside of nested loop
os.walk produces the files and directories as strings, so we have to
cast them into a Path before storing them. The nature of this operation
is a nested loop: iterate over each dirpath, then iterate over each file
in that dirpath.

Previously, the dirpath was being cast into a Path object inside of the
file loop in several locations.

Cast dirpath above the file loop to eliminate unnecessary casts. This
only matters in cases where a dirpath contained more than one file.