Optimize stage by removing Path.relative_to calls
Path.relative_to() is slow.

Avoid calling it in ModController.stage.

We can avoid calling it because we know how many path.parts we should be
removing from the left of the path. We already know the relevant paths
will share the same prefix, so there's no need for the safety guarantees
of .relative_to (or the slowness of it).