Store case-corrected relative path in mod.files (#82)
This patch is an optimization that improves startup time by roughly 25%
by improving the speed of ModController.stage (and therefor every
command that has to call ModController.stage).

Store case-corrected relative path in mod.files so they doesn't have to
be computed every time we run stage().

This deceptively makes the test suite as a whole a little bit slower
because the tests have to compute the case-corrected destination for
each file even if it doesn't end up being used in a stage command.
Apparently this is a common enough occurrence that it outweighs the
performance benefits from the other areas. Additionally, tests always
start with a clean slate, so they only experience the startup
performance benefits when they run refresh or nest sessions like in
test_controller_instance.