Add tests for custom game selection files
I suspect the 'name' value that I believed belonged in the json files is
actually unused, as the GameSelection.name appears to be derived from
the filename.

I vaguely recall needing to show json['name'] in the game selection UI
at one point, where the actual game that would be managed was
GameSelection.name. I'm not sure how we would show one value in the UI
with another value for the actual game name when the GameSelection
object only has one name value.

Anyway, it appears unused. It's not removed in this patch but at least
there's some tests covering custom game config files now.

Oh, and I moved the custom-game-parsing logic into a dedicated method to
make it easier to test.