Show `find` tutorial when a filter is active
The `find` command might not be as intuitive for others as it is for me.

Add a short tutorial to the top of output when a filter has been applied
with the `find` command. This should help users make the most of this
feature and not get stuck in a state where they don't know how to
display things (and might not read the wall-of-text help menu to figure
it out).