Workaround an issue with mods packaged with duplicate dir names
When mods are packaged with duplicate directory names and the first dir
is the only directory in the mod, previously ammo would try to elevate
the contents of that folder so they're next to the folder itself.

This is fine most of the time, but when mods are packaged with nested
directories that have name conflicts, you'd previously hit:

```
OSError: directory not empty
```

Add a check that the mod is packaged correctly. If it has this issue,
don't prompt the user whether the contents should be elevated, and don't
try to elevate the contents.

Instead, just extract it vanilla style, and warn the user that they need
to manually inspect the mod because it is likely packaged incorrectly.