Add pidfile locking (#85)
Only allow one ammo instance to run at a time.

When ammo starts, expect absence of a pidfile at
/var/run/user/{uid}/ammo/pid, then create that file with the current
python process ID as the contents.

When ammo exits via the exit command, KeyboardInterrupt, EOFError, (or
any type of error, I suppose), delete that pidfile and the ammo folder
it was in.

If ammo is terminated via some other means, like the kill command, that
file will _not_ be cleaned up. In these cases, the user is responsible
for manually deleting the pidfile. Since this could be confusing, the
error message that prints when there's already a running instance of
ammo contains the path to the pidfile and instructs to delete it if
there's no other instance of ammo running.