Use spaces instead of underscores in commands (#86)
At one point in time, commands in ammo were like this:
```
delete mod 0
delete download 0
move plugin 0 1
move mod 0 1
```
But the functions that handled each command (E.g., 'move' or 'delete')
were absolutely massive.

In order to reduce the function size and clean up the code, I'd made a
dedicated function for moving mods, and another dedicated function for
moving plugins, and so on.

The commands became:
```
delete_mod 0
delete_download 0
move_plugin 0 1
move_mod 0 1
```

The code has been much better since then, but the user experience is
worse. I don't want to type underscores in command names.

This patch renames functions like 'do_delete_download' to
'delete_download' so they don't show up to the CLI or help menu. It also
adds disambiguation functions named things like 'do_delete' which calls
the functions specific to deleting different types of things.

This has been done for delete, rename, move, activate, and deactivate.
This has been done for the ModController, the BethesdaController, and
the ToolController.

The result is restoration of the old behavior where we can use spaces in
commands instead of underscores, but the code is much cleaner than it
was last time we could do that.

Also this breaks autocomplete. Will address in a follow-up.

Need to add tests for the disambiguation commands too, which is also not
done.