Add a DownloadController to centralize rename/delete download
The ToolController, the ModController, and the BethesdaController
(through inheritance of the ModController) are each able to rename and
delete downloads.

The logic for this was duplicated between the ToolController and the
ModController. This isn't necessarily horrible, but I kept forgetting to
update the ToolController whenever I made changes to the ModController
for things like sorting downloads or fixing other bugs.

Adding a DownloadController allows the methods for renaming and adding
downloads to be centralized so they only need to be updated in one
place.

Add ammo.controller.download.DownloadController.

It has the same logic for renaming and deleting downloads as
ModController had (as that had less bugs than the ToolController's
implementation).

Make ModController and ToolController inherit DownloadController.

Remove .rename_download and .delete_download implementations from
ModController and ToolController so they rely on the inherited logic.

Update class initialization and __str__ implementations to rely on
inheritance.

Eventually, the 'install' implementation may be moved to the
DownloadController as well.