Change expected behavior when deleting plugin
When a plugin is deleted, we previously expected that plugin to
disappear. That makes the delete plugin command have less utility than
it could.

Instead, only delete the plugin from the winning mod (in cases where
that plugin is provided by multiple mods).

This might be confusing for users when they have plugins provided by
multiple mods, but if they really want to nuke the plugin, they'll now
need to run the delete command for as many times as they have mods which
provide a plugin of that name.