Move Game and BethesdaGame into component.py
I want to move the extract_archive method into the download controller,
because it will make it so I don't have to maintain it in the
ModController and the ToolController.

extract_archive relies on self.game, however, and the download
controller currently has no attribute self.game.

We could give it self.game, but then it would need to import resources
that were in mod.py or bethesda.py, both of which need the
DownloadController class from download.py.

This is a little too close to circular imports for my comfort.

Move Game and BethesdaGame into component.py so that in the future,
download.py can import them from there.