Move extract_archive and install logic into download controller
The extract_archive and do_install methods had the bulk of their logic
duplicated between the ToolController and the ModController.

This was hard to maintain because fixing bugs in the ModController often
didn't result in bugs fixed in the ToolController as well.

Moving the extract_archive and install logic into the DownloadController
allows the ModController and ToolController to vastly simplify their
do_install methods to just:

```
def do_install(self, index: Union[int, str]) -> None:
    return super().install(index, "some directory here")
```