Fix creation kit platform extended being installed to wrong dir
When installing this as a mod (as is likely required, as opposed to
installing it as a tool), it would previously install into self.game_pak
like it were an Unreal mod. This was caused by .pak files being present,
as this was probably made in Unreal. However, we want it to install to
self.game_root instead, otherwise it probably won't be able to work
correctly.

Add an exception for ckpe so it installs to the game root.