Reduce the number of _local._parse_path calls
As the number of mods grows, the number of times we call
pathlib._local._parse_path increase. This call isn't particularly fast,
so any time we can avoid calling it, we get a little speed up.

Add a micro-optimization to reduce the number of times ._parse_path is
called by the number of mods, which is typically going to be less than
0.1% of an improvement; so basically nothing. It is measurable with
cProfile, however.