Make mod files relative to game directory (#95)
Different types of mods expect to have their files put in different
places, and there are multiple ways to package each mod with regards to
how many parent directories (up to and excluding the actual game
directory) the mod contains.

Previously, we were tracking a mod.install_dir which was responsible for
tracking the directory that a mod's files should be extracted to. All of
a mod's files were populated relative to this directory instead of
relative to the game's directory.

This created issues with the 'collisions' command. Namely, the UI would
indicate a mod had collisions with an asterisk, but for some colliding
mods, the collision command would display no output. This could have
been caused by one or more of these unconfirmed behaviors I was
suspicious of the potential existence of:
- File comparison was being done against paths relative to
  mod.install_dir. Mods with different install_dirs would have the bug.
- Casefolding before comparison was being done against paths relative to
  mod.install_dir instead of against paths relative to the game dir.
  Mods might not casefold in a way that resulted in collisions for the
  collisions command.
- Something else I haven't anticipated.

In component.py, make mods populate files relative to the game
directory instead of the mod's install directory.

Remove the mod.install_dir attribute. We still use this data, but it's
contained in the post_init and populate_files logic for this dataclass
now; it doesn't need to be exposed anymore.

In the ModController, make stage() resolve destinations relative to the
game directory instead of each mod's install directory.

This fixes the issue where the collisions command sometimes wouldn't
produce output. It also just makes the files interface more sane.