Ignore case of ignored files (#94)
We have a few files we avoid using to detect collisions and we don't
install them into the game directory.

Previously, the comparison to these files was case-sensitive.

Lowercase any filenames we check for in this list, and declare the
list's files with lowercase characters.

This helps prevent anything getting past the filter, like readMe.txt
from Reduced NPC Headtracking.